refactor: Extractthe PasswordSection component into its own file
